+++ integration/playwright-test/examples/failing/error-thrown.spec.ts
@@ -1,0 +1,11 @@
+import { describe, it } from '@serenity-js/playwright-test';
+
+describe('Playwright Test reporting', () => {
+
+    describe('A scenario', () => {
+
+        it('fails when an error is thrown', () => {
+            throw new Error(`Something happened`); // fail with throw
+        });
+    });
+});