+++ integration/playwright-test/examples/screenplay/assertion-error.spec.ts
@@ -1,0 +1,14 @@
+import { Ensure, equals  } from '@serenity-js/assertions';
+import { describe, it } from '@serenity-js/playwright-test';
+
+describe('Playwright Test reporting', () => {
+
+    describe('A screenplay scenario', () => {
+
+        it('correctly reports assertion errors', async ({ actorCalled }) => {
+            await actorCalled('Donald').attemptsTo(
+                Ensure.that(false, equals(true)),
+            );
+        });
+    });
+});