+++ integration/web-specs/spec/screenplay/interactions/Press.spec.ts
@@ -3,21 +3,19 @@ import 'mocha';
 
 import { expect } from '@integration/testing-tools';
 import { Ensure, equals, matches } from '@serenity-js/assertions';
-import { actorCalled, Check, Log, Question, Task } from '@serenity-js/core';
-import { formatted } from '@serenity-js/core/lib/io';
+import { actorCalled, Check, d, Log, Question, Task } from '@serenity-js/core';
 import { BrowseTheWeb, By, Click, DoubleClick, Enter, Key, Navigate, PageElement, Press, Value } from '@serenity-js/web';
 import { given } from 'mocha-testdata';
 
-/** @test {Press} */
 describe('Press', () => {
 
     const InputBoxForm = {
-        textField:      PageElement.located(By.id('input-box')).describedAs('the text field'),
+        textField: PageElement.located(By.id('input-box')).describedAs('the text field'),
     };
 
     const CopyAndPasteBoxesForm = {
-        source:         PageElement.located(By.id('source')).describedAs('source text field'),
-        destination:    PageElement.located(By.id('destination')).describedAs('destination text field'),
+        source: PageElement.located(By.id('source')).describedAs('source text field'),
+        destination: PageElement.located(By.id('destination')).describedAs('destination text field'),
     };
 
     const OS = () => Question.about('operating system', async actor => {
@@ -27,7 +25,6 @@ describe('Press', () => {
 
     describe('single keys', () => {
 
-        /** @test {Press.the} */
         it('allows the actor to enter keys individually into a field', () =>
             actorCalled('Bernie').attemptsTo(
                 Navigate.to('/screenplay/interactions/press/input_box.html'),
@@ -41,7 +38,6 @@ describe('Press', () => {
 
     describe('key chords', function () {
 
-        /** @test {Press.the} */
         it('allows the actor to use modifier keys', () =>
             actorCalled('Bernie').attemptsTo(
                 Navigate.to('/screenplay/interactions/press/input_box.html'),
@@ -54,7 +50,6 @@ describe('Press', () => {
                 Ensure.that(Value.of(InputBoxForm.textField), equals('hi')),
             ));
 
-        /** @test {Press.the} */
         it('allows the actor to use keyboard shortcuts outside the context of any specific input box', async () => {
 
             const Copy = () => Task.where(`#actor performs a "copy" operation`,
@@ -90,17 +85,16 @@ describe('Press', () => {
             );
         });
 
-        /** @test {Press.the} */
         it('allows the actor to use keyboard shortcuts in the context of a specific input box', async () => {
 
             const SelectValueOf = (field: Question<Promise<PageElement>>) =>
-                Task.where(formatted `#actor selects the value of ${ field }`,
+                Task.where(d `#actor selects the value of ${ field }`,
                     Click.on(field),
-                    DoubleClick.on(field)
+                    DoubleClick.on(field),
                 );
 
             const CopyFrom = (field: Question<Promise<PageElement>>) =>
-                Task.where(formatted `#actor performs a "copy" operation on ${ field }`,
+                Task.where(d `#actor performs a "copy" operation on ${ field }`,
                     SelectValueOf(field),
                     Check.whether(OS(), matches(/^mac|darwin/i))
                         .andIfSo(Press.the(Key.Control, Key.Insert).in(field))
@@ -108,7 +102,7 @@ describe('Press', () => {
                 );
 
             const PasteInto = (field: Question<Promise<PageElement>>) =>
-                Task.where(formatted `#actor performs a "paste" operation on ${ field }`,
+                Task.where(d `#actor performs a "paste" operation on ${ field }`,
                     Check.whether(OS(), matches(/^mac|darwin/i))
                         .andIfSo(Press.the(Key.Shift, Key.Insert).in(field))
                         .otherwise(Press.the(Key.Control, 'v').in(field)),
@@ -117,7 +111,7 @@ describe('Press', () => {
             const LoseFocus = () =>
                 Task.where(`#actor makes sure their browser is not focused on any input box`,
                     Press.the(Key.Escape),
-                )
+                );
 
             await actorCalled('Bernie').attemptsTo(
                 Navigate.to('/screenplay/interactions/press/copy_and_paste_boxes.html'),
@@ -140,45 +134,65 @@ describe('Press', () => {
         {
             description: 'single key, document context',
             interaction: Press.the('a'),
-            expected:   `#actor presses key a`,
+            expected: `#actor presses key a`,
         },
         {
             description: 'single key, element context',
             interaction: Press.the('a').in(InputBoxForm.textField),
-            expected:   `#actor presses key a in the text field`,
+            expected: `#actor presses key a in the text field`,
         },
         {
             description: 'sequence of keys, document context',
             interaction: Press.the('a', 'b', 'c'),
-            expected:   `#actor presses keys a, b, c`,
+            expected: `#actor presses keys a, b, c`,
         },
         {
             description: 'sequence of keys, element context',
             interaction: Press.the('a', 'b', 'c').in(InputBoxForm.textField),
-            expected:   `#actor presses keys a, b, c in the text field`,
+            expected: `#actor presses keys a, b, c in the text field`,
         },
         {
             description: 'keyboard shortcut, document context',
             interaction: Press.the(Key.Meta, 'a'),
-            expected:   `#actor presses keys Meta-a`,
+            expected: `#actor presses keys Meta-a`,
         },
         {
             description: 'keyboard shortcut, element context',
             interaction: Press.the(Key.Meta, 'a').in(InputBoxForm.textField),
-            expected:   `#actor presses keys Meta-a in the text field`,
+            expected: `#actor presses keys Meta-a in the text field`,
         },
         {
             description: 'complex shortcut, document context',
             interaction: Press.the(Key.Meta, Key.Alt, 'a'),
-            expected:   `#actor presses keys Meta-Alt-a`,
+            expected: `#actor presses keys Meta-Alt-a`,
         },
         {
             description: 'complex shortcut, element context',
             interaction: Press.the(Key.Meta, Key.Alt, 'a').in(InputBoxForm.textField),
-            expected:   `#actor presses keys Meta-Alt-a in the text field`,
+            expected: `#actor presses keys Meta-Alt-a in the text field`,
         },
     ]).
     it('provides a sensible description of the interaction being performed', ({ interaction, expected }) => {
         expect(interaction.toString()).to.equal(expected);
+    });
+
+    describe('detecting invocation location', () => {
+        it('correctly detects its invocation location', () => {
+            const activity = Press.the('a');
+            const location = activity.instantiationLocation();
+
+            expect(location.path.basename()).to.equal('Press.spec.ts');
+            expect(location.line).to.equal(181);
+            expect(location.column).to.equal(36);
+        });
+
+        it('correctly detects its invocation location when custom errors are used', () => {
+            const activity = Press.the('a').in(InputBoxForm.textField);
+            const location = activity.instantiationLocation();
+
+            expect(location.path.basename()).to.equal('Press.spec.ts');
+            expect(location.line).to.equal(190);
+            expect(location.column).to.equal(47);
+        });
     });
 });