+++ packages/core/spec/io/loader/examples/Example.ts
@@ -1,0 +1,4 @@
+export class Example {
+    constructor(public readonly name: string) {
+    }
+}