+++ packages/core/spec/io/loader/examples/ExampleWithFactoryMethod.ts
@@ -1,0 +1,8 @@
+export class ExampleWithFactoryMethod {
+    static fromJSON(value: { name: string }): ExampleWithFactoryMethod {
+        return new ExampleWithFactoryMethod(value.name)
+    }
+
+    constructor(public readonly name: string) {
+    }
+}