+++ packages/core/src/index.ts
@@ -1,8 +1,8 @@
+export * from './config';
 export * from './errors';
 export * from './instance';
 export { d, f, format } from './io';
 export { Duration, Timestamp } from './model';
 export * from './screenplay';
 export * from './Serenity';
-export { SerenityConfig } from './SerenityConfig';
 export * from './stage';