+++ packages/core/src/io/loader/index.ts
@@ -1,0 +1,5 @@
+export * from './ClassDescriptionParser';
+export * from './ClassDescriptor';
+export * from './ClassLoader';
+export * from './ModuleLoader';
+export * from './Version';